plotFun() now handles limit setting better
For any variable v in rhs of the formula, vlim and v.lim can be used to set the limits.  The use of v= has been removed since it is a poor choice of name and conflicts with other things (like the use of x as the first argument in lattice functions).