giving plotFun more access to lattice defaults
Some refactoring work to make plotFun() behave more like other lattice
functions and to use default settings from trellis.par.get().

The basic strategy has been to push things into ... but to
manipulate ... a bit so that it can still also be used to set free
variables in the function being plotted:

	funPlot( sin(a*x) ~ x, a=3 )

Also doing some renaming to make variable names and usage align better
with lattice graphics.  (Examples: alpha, col.regions.)

col.regions can now be either a function or a vector of colors.  A
utility function will convert a vector of colors into a function that
returns a vector of colors of the desired length and adds in alpha
values.