groups now "working" in plotFun()
The implementation is not perfect, but it does something reasonable.  Note: expressions defining groups can be defined in terms of names used in the expression being plotted or in terms of canonical names (x,y, and .height).  Could choose a different canonical name for the height in a 2d graph perhaps.

Documentation needs to be updated to describe groups argument.